requirements.txt
Django==2.1.15
django-ckeditor==5.8.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-notifications-hq==1.5.0
greenlet==0.4.15
jsonfield==2.0.2
Pillow==6.2.1
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
